[refactor/#217] setting 패키지 수정
